Relocate some code from controller to service.
